backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
aiosqlite
Jinja2
python-multipart
email-validator
python-dotenv
Pillow
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
websockets
aiortc
pywebpush
python-magic
aiofiles
ffmpeg-python
fastapi-utils
requests
httpx
reportlab
matplotlib
anthropic
